[Update] 拆分profile view
