+++ apps/users/templates/users/_base_otp.html
@@ -10,7 +10,7 @@
                 {% endblock %}
             </h2>
         </div>
-        <div >
+        <div>
             <div class="verify">{% trans 'Security token validation' %}&nbsp;&nbsp;{% trans 'Account' %}&nbsp;<span>{{ user.username }}</span>&nbsp;&nbsp;{% trans 'Follow these steps to complete the binding operation' %}</div>
             <hr style="width: 500px; margin: auto; margin-top: 10px;">
             {% block content %}
