+++ apps/users/templates/users/forgot_password.html
@@ -3,11 +3,6 @@
 {% load i18n %}
 {% load bootstrap3 %}
 {% block custom_head_css_js %}
-    <style>
-        .captcha {
-             float: right;
-        }
-    </style>
 {% endblock %}
 {% block html_title %}{% trans 'Forgot password' %}{% endblock %}
 {% block title %} {% trans 'Forgot password' %}?{% endblock %}
