+++ apps/users/templates/users/user_otp_check_password.html
@@ -10,15 +10,11 @@
     <form class="" role="form" method="post" action="">
         {% csrf_token %}
         <div class="form-input">
-            <input type="password" class="" name="{{ form.password.html_name }}" placeholder="{% trans 'Password' %}" required="">
+            <input type="password" class="" name="{{ form.password.html_name }}" placeholder="{% trans 'Password' %}" required="" autofocus="autofocus">
         </div>
-
         <button type="submit" class="next">{% trans 'Next' %}</button>
-
         {% if 'password' in form.errors %}
             <p class="red-fonts">{{ form.password.errors.as_text }}</p>
         {% endif %}
-
     </form>
 {% endblock %}
-
