+++ apps/users/templates/users/user_profile.html
@@ -166,7 +166,7 @@
                                                                 ">{% trans 'Disable' %}
                                                             {% endif %}
                                                         {% else %}
-                                                            {% url 'users:user-verify-password' %}
+                                                            {% url 'users:user-otp-enable-start' %}
                                                             ">{% trans 'Enable' %}
                                                         {% endif %}
                                                 </a>
