+++ apps/users/views/profile/__init__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+#
+from .base import *
+from .password import *
+from .pubkey import *
+from .mfa import *
+from .otp import *
