+++ apps/users/views/profile/mfa.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+#
